photoresistor-2.007
Using a photoresistor to sense light, and output values in the serial port at 9600 baud setting in the lower right hand corner. 
Connect sig 5V to side 1 of the photoresistor, then side 2 of the photoresistor to an analog pin (e.g. A4), and side 2 also to a resistor. The other end of the resistor then goes to ground. When light is applied, the photoresistor decreases its resistance, increasing the voltage output into the analog A4 pin!