potentiometer-controls-servos-2.007-Ver1
potentiometer controls 2 servos. When light is applied to the potentiometer, it senses the light, decreases in resistance, and thus provides higher voltage to the analog pin. Via an "if" statement in the loop, I can then control depending on the light exposure, whether or not the servos will engage and go through a predetermined set of motions.